Restructure keystone package
