+++ .changeset/06b7f52d/changes.md
@@ -1,0 +1,1 @@
+- Use named exports from @keystone-alpha/keystone package.
