+++ .changeset/894c5860/changes.md
@@ -1,0 +1,1 @@
+- Make all parts of the API available as named exports.
