+++ packages/keystone/index.js
@@ -1,1 +1,15 @@
-exports.Keystone = require('./Keystone');
+const Keystone = require('./lib/Keystone');
+const PasswordAuthStrategy = require('./lib/auth/Password');
+const TwitterAuthStrategy = require('./lib/auth/Twitter');
+const { FacebookAuthStrategy } = require('./lib/auth/Facebook');
+const { BaseKeystoneAdapter, BaseListAdapter, BaseFieldAdapter } = require('./lib/adapters');
+
+module.exports = {
+  Keystone,
+  PasswordAuthStrategy,
+  TwitterAuthStrategy,
+  FacebookAuthStrategy,
+  BaseKeystoneAdapter,
+  BaseListAdapter,
+  BaseFieldAdapter,
+};
